requirements.txt
# Core Dependencies
python-dotenv==1.0.0
requests==2.32.3
pydantic>=2.7.2

# Google AI and Workspace
google-genai==1.51.0
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Web Framework for Webhooks
fastapi==0.109.0
uvicorn[standard]==0.26.0
python-multipart==0.0.6

# Optional but recommended
httpx>=0.27.0
aiofiles==23.2.1

# For memory embeddings and vector store
numpy>=1.24.0

# Date parsing
python-dateutil>=2.8.0

# Performance profiling and monitoring
memory-profiler>=0.61.0
psutil>=5.9.0

# Load testing
locust>=2.15.0

# MCP
mcp>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-cov>=4.1.0
coverage>=7.3.0
responses>=0.24.0
freezegun>=1.2.0
aiohttp>=3.9.0
